Use separate permutation tables for each coordinate
This is based on a technique described in the paper Better Gradient Noise.
Instead of P[P[P[P[x] + y] + z] + w] for a lookup we instead make four
permutation tables and do Px[x] ^ Py[y] ^ Pz[z] ^ Pw[w]. This not only
improves performance by reducing data dependencies it also gives a better
output by decorrelating the indices. As we no longer need to double the size
of the tables this only doubles the number of entries we need to store.
After switching the tables to u8s this should actually only use half the
memory on 32-bit systems and one quarter the memory on 64-bit systems.