Change Seed back to one permutation table
The changes in commit ce8e06ff made Seed use multiple permutation tables
which resulted in a significant speed improvement. However, when using
multiple tables in this style you are required to use a much larger
gradient/vector table in order to get correct results. The minimum number
of entries required is 64 while the ideal is 256. Doing this changes the
time for a single perlin2 sample from 13ns to 4000ns which is not
acceptable.

As we can not properly make use of the multiple permutation tables we
instead swhich back to only use one. However we are now using proper
Pearson hashing instead of the variant we had before. This allows us to
avoid having a double length table and seems to still offer a small benefit
in speed over the old style, even if it isn't as large as multiple tables
seemed to be.