Make PermutationTable public
This allows the noise functions in core to be called from outside the
library.

Also update the perlin example to call the noise functions in core
instead of using the Perlin NoiseFn struct. Required a sort of shim to
make the PlaneMapBuilder work with a function pointer instead of a
NoiseFn.