Remove simplex_1d function
A 2D simplex is a equilateral triangle (3 sides), and a 3D simplex is a tetrahedron (4 faces).
2D Perlin uses a square (4 sides), and 3D Perlin uses a cube (8 faces).

A 1D simplex is a line, the same as what 1D Perlin noise uses. As such, 1D simplex noise is equivalent to 1D Perlin noise. Hence, removing the completely redundant 1D Simplex noise function.